1.1.0: Make the app compatible with the 2.13+ players.
Starting with 2.13, the player will pass a DOM container to the apps
during the prepare() call. The apps are expected to work on that
container instead of working on document.body or their own container.

The application no longer needs to clean up the DOM nodes from a
previous view. The player will automatically clean up the provided
container when the view is done.

In addition to the done() callback, the player will provide a begin()
callback when executing a view. The application needs to call begin()
whenever it finishes modifying the scene.